LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/dashed-function-cycles-expected.txt


FAIL Local with self-cycle assert_equals: expected "PASS" but got "var(--x, PASS)"
FAIL Cycle reference without fallback makes result invalid assert_equals: expected "" but got "var(--x)"
FAIL Local with self-cycle in unused fallback assert_equals: expected "PASS" but got "var(--x, FAIL)"
FAIL Local shadowing cyclic property --x assert_equals: expected "PASS" but got "var(--x)"
FAIL Local shadowing cyclic outer local --x assert_equals: expected "PASS" but got "--g()"
FAIL Argument shadowing cyclic outer local --x assert_equals: expected "10px" but got "--g(10px)"
FAIL Arguments shadowing cyclic properties assert_equals: expected "PASS-x PASS-y" but got "var(--x) var(--y)"
FAIL Observing property cycle locally assert_equals: expected "PASS-x PASS-y" but got "var(--x, PASS-x) var(--x, PASS-y)"
FAIL Using cyclic values with no fallback assert_equals: expected "" but got "var(--x) var(--y)"
PASS Self-cycle in non-used local variable
FAIL Using cyclic value in unused fallback assert_equals: expected "PASS" but got "var(--x, var(--y))"
FAIL Using cyclic value in unused fallback (local) assert_equals: expected "PASS" but got "var(--x, var(--y))"
FAIL Dashed-function, self-cycle assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through other function (--g) assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through other function (--f) assert_equals: expected "PASS" but got "--g()"
FAIL Cycle through local, self assert_equals: expected "PASS" but got "var(--local)"
FAIL Cycle through unused local assert_equals: expected "PASS" but got "FAIL-result"
FAIL Cycle through global, self assert_equals: expected "PASS" but got "var(--global)"
FAIL Cycle through local, other function assert_equals: expected "PASS" but got "var(--local)"
FAIL Cycle through local, other function, fallback in function assert_equals: expected "PASS" but got "var(--a, PASS)"
FAIL Cycle through various variables and other functions assert_equals: expected "PASS" but got "var(--global)"
FAIL Function in a cycle with its own default assert_equals: expected "PASS" but got "10px"
FAIL Cyclic defaults assert_equals: expected "42px PASS-y PASS-z" but got "var(--x, FAIL) var(--y, PASS-y) var(--z, PASS-z)"
FAIL Cyclic outer --b shadows custom property assert_equals: expected "PASS" but got "var(--a)"
FAIL Locals are function specific assert_equals: expected "10px" but got "var(--a)"
